requirements.txt
demucs==4.0.1
einops==0.8.1
fast_bss_eval==0.1.4
mamba_ssm==2.2.5
mir_eval==0.8.2
numpy==2.2.6
pandas==2.3.1
PyYAML==6.0.2
PyYAML==6.0.2
soundfile==0.13.1
swanlab==0.6.8
tensorboardx==2.6.4
thop==0.1.1.post2209072238
timm==1.0.19
torchaudio==2.7.0
tqdm==4.67.1
